requirements.txt
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.0
anyio==3.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
crcmod==1.7
cryptography==40.0.2
dnspython==2.3.0
fastapi==0.95.2
filelock==3.12.0
fsspec==2023.5.0
h11==0.14.0
httptools==0.5.0
huggingface-hub==0.14.1
idna==3.4
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
oss2==2.17.0
packaging==23.1
Pillow==9.5.0
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
pymongo==4.3.3
python-dotenv==1.0.0
PyYAML==6.0
redis==4.5.5
regex==2023.5.5
requests==2.30.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.29.2
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
